+++ src/sources/syslog/factory.rs
@@ -66,9 +66,10 @@ impl SourceFactory for SyslogSourceFactory {
             let svc = match config.protocol {
                 Protocol::Udp => {
                     info_ctrl!(
-                        "syslog UDP factory build: strip_header={}, attach_meta_tags={}, udp_recv_buffer={}",
+                        "syslog UDP factory build: strip_header={}, attach_meta_tags={}, fast_strip={}, udp_recv_buffer={}",
                         config.strip_header,
                         config.attach_meta_tags,
+                        config.fast_strip,
                         config.udp_recv_buffer
                     );
                     let meta = meta_builder(&tags);
@@ -78,6 +79,7 @@ impl SourceFactory for SyslogSourceFactory {
                         tags,
                         config.strip_header,
                         config.attach_meta_tags,
+                        config.fast_strip,
                         config.udp_recv_buffer,
                     )
                     .await?;